+++ 1-stdlib/deferred/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## PENDING
 * [chore] [breaking] cjs pre-build now targeting node12/ES2019 [details](../../CONTRIBUTING/module-exports.md)
+* [feat] slightly improved typings, following the latest Promise definition (Typescript 3.8.3)
 * [doc] credits
 * ...