+++ 1-stdlib/deferred/CHANGELOG.md
@@ -2,12 +2,15 @@
 **This package follows [semver](https://semver.org/).**
 
 ## PENDING
+* TODO unit tests!
+* ...
+
+## 3.0.0
+2020/06/30
 * [chore] [breaking] will now by default prevent the "uncaught rejection" message
 * [chore] [breaking] cjs pre-build now targeting node12/ES2019 [details](../../CONTRIBUTING/module-exports.md)
 * [feat] slightly improved typings, following the latest Promise definition (Typescript 3.8.3)
 * [doc] credits
-* TODO unit tests!
-* ...
 
 ## 2.0.1
 2019/12/12